+++ src/string.c
@@ -15,15 +15,14 @@ int bka_strlen (char const *str) {
 }
 
 
-int bka_atoi10(char const *src){
+long int bka_atoi10 (char const *src) {
 	int i = 0;
 	int res = 0;
 	
-	while(src[i] != '\0')
-			{
-				res = res*10 + (src[i] - '0');
-				i+=1;
-			}
+	while (src[i] != '\0') {
+		res = res * 10 + (src[i] - '0');
+		i += 1;
+	}
 			
 	return res;
 }
